requirements.txt
dbt-core
dbt-snowflake
apache-airflow-providers-snowflake
git